+++ apps/src/code-studio/pd/professional_learning/plcElements/ModuleAssignment.story.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ModuleAssignment from '@cdo/apps/code-studio/pd/professional_learning_landing/plcElements/ModuleAssignment';
+import ModuleAssignment from '@cdo/apps/code-studio/pd/professional_learning/plcElements/ModuleAssignment';
 
 export default {
   component: ModuleAssignment,
