+++ apps/src/code-studio/pd/professional_learning/tableStyles.scss
