+++ apps/src/sites/studio/pages/pd/professional_learning_landing/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 
 import {displayDifferentiationChat} from '@cdo/apps/aiDifferentiation/aiDiffUtils';
-import LandingPage from '@cdo/apps/code-studio/pd/professional_learning_landing/LandingPage';
+import LandingPage from '@cdo/apps/code-studio/pd/professional_learning/LandingPage';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import analyticsReporter from '@cdo/apps/metrics/AnalyticsReporter';
 import {getStore} from '@cdo/apps/redux';
