+++ apps/src/sites/studio/pages/plc/user_course_enrollments/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import ProfessionalLearningCourseProgress from '@cdo/apps/code-studio/pd/professional_learning_landing/ProfessionalLearningCourseProgress';
+import ProfessionalLearningCourseProgress from '@cdo/apps/code-studio/pd/professional_learning/ProfessionalLearningCourseProgress';
 import getScriptData from '@cdo/apps/util/getScriptData';
 
 const userCourseEnrollmentData = getScriptData('userCourseEnrollmentData');
